+++ pkg/NEWS.md
@@ -1,10 +1,12 @@
-# Version 0.5-6 (development, rev. 641-650)
+# Version 0.5-7 (development, rev. 641-650)
 
 * adapt to OverlayNG in GEOS 3.9
 
 * update Windows static library to GEOS 3.9
 
 * add gMinumumRotatedRectangle() and gMaximumInscribedCircle() for GEOS 3.9
+
+* adapt configure.ac for GEOS 3.10
 
 # Version 0.5-5 (2020-09-07, rev. 634-640)